+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+-  AthleticsHomePanel update semantics label for image [#510](https://github.com/rokwire/illinois-app/issues/510).
 
 ## [2.6.22] - 2021-08-23
 ### Fixed