+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.0.2] 2021-08-27
 ### Fixed
--  AthleticsHomePanel update semantics label for image [#510](https://github.com/rokwire/illinois-app/issues/510).
+- AthleticsHomePanel update semantics label for image [#510](https://github.com/rokwire/illinois-app/issues/510).
 - BrowsePanel: improve Large Text support [#511](https://github.com/rokwire/illinois-app/issues/511).
 ### Added
 - Created initial Inbox functionality and UI [#714](https://github.com/rokwire/illinois-app/issues/714).