+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Fixed
+- Remove Converge url action [#292](https://github.com/rokwire/illinois-app/issues/292).
 -  AthleticsHomePanel update semantics label for image [#510](https://github.com/rokwire/illinois-app/issues/510).
 - BrowsePanel: improve Large Text support [#511](https://github.com/rokwire/illinois-app/issues/511).